Lab10/hh_back/requirements.txt
Django~=3.2.10
djangorestframework~=3.14.0